Move domain block receipt protocol to domain network
