+++ domains/client/domain-operator/src/lib.rs
@@ -83,11 +83,13 @@ pub use self::aux_schema::load_execution_receipt;
 pub use self::fetch_domain_bootstrap_info::{fetch_domain_bootstrap_info, BootstrapResult};
 pub use self::operator::Operator;
 pub use self::utils::{DomainBlockImportNotification, DomainImportNotifications, OperatorSlotInfo};
+pub use crate::snap_sync::LastDomainBlockReceiptProvider;
 pub use domain_worker::OpaqueBundleFor;
 use futures::channel::mpsc;
 use futures::Stream;
 use sc_client_api::{AuxStore, BlockImportNotification};
 use sc_consensus::BoxBlockImport;
+use sc_network::service::traits::NetworkService;
 use sc_network::NetworkRequest;
 use sc_network_sync::block_relay_protocol::BlockDownloader;
 use sc_network_sync::service::network::NetworkServiceHandle;
@@ -161,7 +163,6 @@ pub struct OperatorParams<
     CIBNS,
     NSNS,
     ASS,
-    NR,
     CNR,
 > where
     Block: BlockT,
@@ -170,7 +171,6 @@ pub struct OperatorParams<
     CIBNS: Stream<Item = BlockImportNotification<CBlock>> + Send + 'static,
     NSNS: Stream<Item = NewSlotNotification> + Send + 'static,
     ASS: Stream<Item = mpsc::Sender<()>> + Send + 'static,
-    NR: NetworkRequest + Send + Sync,
     CNR: NetworkRequest + Send + Sync + 'static,
 {
     pub domain_id: DomainId,
@@ -191,11 +191,12 @@ pub struct OperatorParams<
     pub skip_empty_bundle_production: bool,
     pub skip_out_of_order_slot: bool,
     pub sync_service: Arc<SyncingService<Block>>,
-    pub network_request: NR,
+    pub network_service: Arc<dyn NetworkService>,
     pub block_downloader: Arc<dyn BlockDownloader<Block>>,
-    pub consensus_chain_sync_params: Option<ConsensusChainSyncParams<Block, CBlock, CNR>>,
+    pub consensus_chain_sync_params: Option<ConsensusChainSyncParams<CBlock, CNR>>,
     pub domain_fork_id: Option<String>,
     pub domain_network_service_handle: NetworkServiceHandle,
+    pub domain_execution_receipt_provider: Arc<dyn LastDomainBlockReceiptProvider<Block, CBlock>>,
 }
 
 pub(crate) fn load_execution_receipt_by_domain_hash<Block, CBlock, Client>(
