+++ domains/test/service/src/domain.rs
@@ -217,7 +217,7 @@ where
             maybe_operator_id,
             confirmation_depth_k: chain_constants.confirmation_depth_k(),
             consensus_chain_sync_params: None::<
-                ConsensusChainSyncParams<_, _, Arc<dyn NetworkRequest + Sync + Send>>,
+                ConsensusChainSyncParams<_, Arc<dyn NetworkRequest + Sync + Send>>,
             >,
         };
